requirements.txt
algoliasearch-django>=2.0,<3.0
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.9.14
charset-normalizer==2.1.1
Django==4.0.7
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt
idna==3.4
pycodestyle==2.9.1
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.12
